# api/openapi.yaml
openapi: 3.1.0
info:
  title: Neurun API
  version: 0.1.0
  description: Build Python deployments and execute immutable builds.
servers:
  - url: http://localhost:1267
security:
  - bearerAuth: []
paths:
...
  /v1/projects:
    get:
      operationId: listProjects
      parameters:
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Projects visible to the principal.
          content:
            application/json:
              schema:
                type: object
                required: [projects]
                properties:
                  projects:
                    type: array
                    items: {$ref: "#/components/schemas/Project"}
  /v1/projects/{project_id}:
    parameters:
      - name: project_id
        in: path
        required: true
        schema: {type: string}
    get:
      operationId: getProject
      responses:
        "200":
          description: Project.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Project"}
        "404": {$ref: "#/components/responses/Problem"}
    patch:
      operationId: updateProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [name]
              properties:
                name: {type: string}
      responses:
        "200":
          description: Updated project.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Project"}
  /v1/apps:
    get:
      operationId: listApps
      parameters:
        - $ref: "#/components/parameters/Limit"
        - name: name
          in: query
          description: Return the app with this exact project-scoped name.
          schema: {type: string, minLength: 1, maxLength: 120}
      responses:
        "200":
          description: Apps in the authenticated project.
          content:
            application/json:
              schema:
                type: object
                required: [apps]
                properties:
                  apps:
                    type: array
                    items: {$ref: "#/components/schemas/App"}
        "422": {$ref: "#/components/responses/Problem"}
    post:
      operationId: createApp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [name, repository]
              properties:
                name: {type: string}
                repository:
                  type: string
                  description: >-
                    owner/name on GitHub. Required: an app has no other source
                    of code, and the installation must already be able to read
                    it.
                production_ref:
                  type: string
                  description: >-
                    The ref whose pushes deploy. Empty follows the repository's
                    default branch.
      responses:
        "201":
          description: App created.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/App"}
        "409": {$ref: "#/components/responses/Problem"}
  /v1/apps/{app_id}:
    parameters:
      - name: app_id
        in: path
        required: true
        schema: {type: string}
    get:
      operationId: getApp
      responses:
        "200":
          description: App.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/App"}
        "404": {$ref: "#/components/responses/Problem"}
    patch:
      operationId: updateApp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [name]
              properties:
                name: {type: string}
      responses:
        "200":
          description: App updated.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/App"}
        "404": {$ref: "#/components/responses/Problem"}
        "409": {$ref: "#/components/responses/Problem"}
  /v1/apps/{app_id}/repository:
    parameters:
      - name: app_id
        in: path
        required: true
        schema: {type: string}
    put:
      operationId: connectRepository
      description: >-
        Point the app at a GitHub repository, or disconnect it by sending an
        empty repository. The installation must already be able to read it: the
        ref is resolved before anything is stored.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [repository]
              properties:
                repository:
                  type: string
                  description: owner/name, or empty to disconnect.
                production_ref:
                  type: string
                  description: >-
                    The ref a push must touch to deploy. Empty follows the
                    repository's default branch.
      responses:
        "200":
          description: App connected.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/App"}
        "403": {$ref: "#/components/responses/Problem"}
        "404": {$ref: "#/components/responses/Problem"}
        "409": {$ref: "#/components/responses/Problem"}
  /v1/github/installation:
    get:
      operationId: getInstallation
      responses:
        "200":
          description: The organization's GitHub App installation.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Installation"}
        "409": {$ref: "#/components/responses/Problem"}
        "503": {$ref: "#/components/responses/Problem"}
    post:
      operationId: recordInstallation
      description: >-
        Record the installation GitHub redirects back with after somebody
        installs the app. The account it is installed on is read from GitHub,
        since the redirect carries only the installation id. Re-installing
        replaces the previous record.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [installation_id]
              properties:
                installation_id: {type: string}
      responses:
        "201":
          description: Installation recorded.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Installation"}
        "409": {$ref: "#/components/responses/Problem"}
        "422": {$ref: "#/components/responses/Problem"}
        "503": {$ref: "#/components/responses/Problem"}
    delete:
      operationId: deleteInstallation
      responses:
        "204": {description: Installation forgotten.}
        "409": {$ref: "#/components/responses/Problem"}
  /v1/github/repositories:
    get:
      operationId: listRepositories
      description: The repositories this organization's installation can read.
      responses:
        "200":
          description: Repositories the installation grants.
          content:
            application/json:
              schema:
                type: object
                required: [repositories]
                properties:
                  repositories:
                    type: array
                    items: {$ref: "#/components/schemas/Repository"}
        "409": {$ref: "#/components/responses/Problem"}
        "503": {$ref: "#/components/responses/Problem"}
  /v1/github/branches:
    get:
      operationId: listBranches
      parameters:
        - name: repository
          in: query
          required: true
          description: owner/name.
          schema: {type: string}
      responses:
        "200":
          description: Branch names.
          content:
            application/json:
              schema:
                type: object
                required: [branches]
                properties:
                  branches:
                    type: array
                    items: {type: string}
        "403": {$ref: "#/components/responses/Problem"}
        "404": {$ref: "#/components/responses/Problem"}
        "409": {$ref: "#/components/responses/Problem"}
        "503": {$ref: "#/components/responses/Problem"}
  /v1/github/deployments:
    post:
      operationId: deployRef
      description: >-
        Build a ref of the app's connected repository. An absent ref uses the
        app's production ref. A push to that ref does the same thing without
        this call.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [app_id]
              properties:
                app_id: {type: string}
                ref: {type: string}
      responses:
        "201":
          description: Deployment built.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Deployment"}
        "403": {$ref: "#/components/responses/Problem"}
        "404": {$ref: "#/components/responses/Problem"}
        "409": {$ref: "#/components/responses/Problem"}
        "413": {$ref: "#/components/responses/Problem"}
        "503": {$ref: "#/components/responses/Problem"}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    Limit:
      name: limit
      in: query
      schema: {type: integer, minimum: 1, maximum: 200, default: 50}
...
  responses:
    Problem:
      description: Request could not be completed.
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ErrorEnvelope"}
  schemas:
...
    Project:
      type: object
      required: [id, organization_id, name, created_at, updated_at]
      properties:
        id: {type: string}
        organization_id: {type: string}
        name: {type: string}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
    App:
      type: object
      required: [id, project_id, name, created_at, updated_at]
      properties:
        id: {type: string}
        project_id: {type: string}
        name: {type: string}
        repository:
          type: string
          description: owner/name on GitHub. Absent when the app is not connected.
        production_ref:
          type: string
          description: The ref whose pushes deploy. Absent follows the default branch.
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
    Repository:
      type: object
      required: [full_name, default_branch, private]
      properties:
        full_name: {type: string}
        default_branch: {type: string}
        private: {type: boolean}
    Installation:
      type: object
      required: [id, organization_id, installation_id, account_login, created_at, updated_at]
      properties:
        id: {type: string}
        organization_id: {type: string}
        installation_id: {type: integer, format: int64}
        account_login: {type: string}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
...
    Artifact:
      type: object
      required: [id, kind, name, media_type, size_bytes, sha256, created_at]
      properties:
        id: {type: string}
        kind: {type: string}
        name: {type: string}
        media_type: {type: string}
        size_bytes: {type: integer, format: int64}
        sha256: {type: string}
        created_at: {type: string, format: date-time}
    Failure:
      type: object
      required: [code, message]
      properties:
        code: {type: string}
        message: {type: string}
    Build:
      description: >-
        What a deployment produced: the artifacts, and what identifies them. A
        build carries no status and no failure — the deployment that ran it
        does, and a deployment that produced nothing has no build at all.
      type: object
      required: [id, runtime, entrypoint, source_sha256, artifacts, created_at]
      properties:
        id: {type: string}
        runtime: {type: string, enum: [python, rust, go, ruby, node]}
        entrypoint: {type: string}
        source_sha256: {type: string}
        artifacts:
          type: array
          items: {$ref: "#/components/schemas/Artifact"}
        created_at: {type: string, format: date-time}
    Deployment:
      description: >-
        One attempt to turn a source archive into a build. Statuses run
        queued → building → publishing → ready or failed; logs carry what the
        toolchain printed and arrive while it is still running.
      type: object
      required: [id, project_id, app_id, runtime, entrypoint, status, source, logs, created_at, updated_at]
      properties:
        id: {type: string}
        project_id: {type: string}
        app_id: {type: string}
        runtime: {type: string, enum: [python, rust, go, ruby, node]}
        entrypoint: {type: string}
        status: {type: string, enum: [queued, building, publishing, ready, failed]}
        source: {$ref: "#/components/schemas/Artifact"}
        commit_sha: {type: string}
        git_ref: {type: string}
        build:
          oneOf:
            - $ref: "#/components/schemas/Build"
            - type: "null"
        failure:
          oneOf:
            - $ref: "#/components/schemas/Failure"
            - type: "null"
        logs: {type: string}
        started_at:
          type: [string, "null"]
          format: date-time
        finished_at:
          type: [string, "null"]
          format: date-time
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
...
    Problem:
      type: object
      required: [code, message]
      properties:
        code: {type: string}
        message: {type: string}
        details: {type: object, additionalProperties: true}
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error: {$ref: "#/components/schemas/Problem"}